% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_forecasts.R
\name{averaged_forecast}
\alias{averaged_forecast}
\title{Averaged forecast}
\usage{
averaged_forecast(history, start, horizon = 1, average_period = 3)
}
\arguments{
\item{history}{Data.frame. Rows are timestamps, columns are power lines.}

\item{start}{Integer. Unix timestamp of the first forecast time.}

\item{horizon}{Integer. Number of timesteps to forecast.}

\item{average_period}{Integer. Number of last observations for which values should be averaged.}
}
\value{
Data.frame. Rows are timestamps, columns are power lines.
}
\description{
Takes the average of the last average_period many measured value and use them as a constant forecast.
}
