% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_data_layer.R
\name{convert_and_save_power_lines}
\alias{convert_and_save_power_lines}
\title{This function reads target data (power line measurements) in HDF5 format and saves it in RDS format.}
\usage{
convert_and_save_power_lines(ids = c(), dataset = "train",
  target = "/home/datascience/h2020_grid_forecast/R_data/",
  designation = "")
}
\arguments{
\item{ids}{Vector. Defaults to an empty vector which means that all available power lines are selectd. 
Can designate selected power lines by their ids.}

\item{dataset}{String. Indicates an element in the config file's \code{datasets} list.}

\item{target}{String. Folder where to save the file.}

\item{desgination}{String. File name.}
}
\description{
This function reads target data (power line measurements) in HDF5 format and saves it in RDS format.
}
\author{
Jakob Etzel
}
