% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_forecasts.R
\name{copy_last_day}
\alias{copy_last_day}
\title{Forecast with last day's values}
\usage{
copy_last_day(history, start, horizon = 1)
}
\arguments{
\item{history}{Data.frame. Rows are timestamps, columns are power lines.}

\item{start}{Integer. Unix timestamp of the first forecast time.}

\item{horizon}{Integer. Number of timesteps to forecast.}
}
\value{
Data.frame. Rows are timestamps, columns are power lines.
}
\description{
Forecast with last day's values
}
