% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_data_layer.R
\name{get_gosat}
\alias{get_gosat}
\title{This function reads GOSAT data in HDF5 format.}
\usage{
get_gosat(dataset = "aux_gosat_adapt", folders = c())
}
\arguments{
\item{dataset}{String. Indicates an element in the config file's \code{datasets} list.}

\item{folders}{Vector. Defaults to an empty vector which means that all available folders being selectd. 
Can designate selected folders.}
}
\value{
A list of three data.frame objects called CO2, CH4 and H2O.
Each data.frame has a column 'timestamp' and different measurement and metadata columns.
}
\description{
This function reads GOSAT data in HDF5 format.
}
\author{
Jakob Etzel
}
