% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_data_layer.R
\name{get_opsd_not_NaN}
\alias{get_opsd_not_NaN}
\title{This function reads OPSD data in HDF5 format and keeps only non-empty columns.}
\usage{
get_opsd_not_NaN(dataset = "aux_opsd_15_train")
}
\arguments{
\item{dataset}{String. Indicates an element in the config file's \code{datasets} list.}
}
\value{
Data.frame with all measurement columns and a column timestamp with UNIX epoch timestamps.
}
\description{
A column is non-empty if at least one value is not \code{NaN}.
}
\author{
Jakob Etzel
}
