% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_forecasts.R
\name{persistence_forecast}
\alias{persistence_forecast}
\title{Persistence forecast}
\usage{
persistence_forecast(history, start, horizon = 1)
}
\arguments{
\item{history}{Data.frame. Rows are timestamps, columns are power lines.}

\item{start}{Integer. Unix timestamp of the first forecast time.}

\item{horizon}{Integer. Number of timesteps to forecast.}
}
\value{
Data.frame. Rows are timestamps, columns are power lines.
}
\description{
Takes the last measured value and use them as a constant forecast.
}
