% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{rmse}
\alias{rmse}
\title{Evaluates forecasts with RMSE}
\usage{
rmse(forecast, actual)
}
\arguments{
\item{forecast}{Data.frame. Rows are timestamps, columns are power lines.}

\item{actual}{Data.frame. Rows are timestamps, columns are power lines.}
}
\value{
Double. RMSE over all forecast steps, horizons and power lines.
}
\description{
The RMSE is the root mean square error. The function tests whether dimensions and 
timestamps are equal for forecast and actual. If this is not the case, NaN will be
returned
}
