% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_data_layer.R
\name{get_opsd}
\alias{get_opsd}
\title{This function reads OPSD data in HDF5 format.}
\usage{
get_opsd(dataset = "aux_opsd_15_train")
}
\arguments{
\item{dataset}{String. Indicates an element in the config file's \code{datasets} list.}
}
\value{
Data.frame with all measurement columns and a column timestamp with UNIX epoch timestamps.
}
\description{
This function reads OPSD data in HDF5 format.
}
\author{
Jakob Etzel
}
